refactor(kubernetes): Move a few classes to better packages
KubernetesV2SearchProvider should be in the same package as all
the other Kubernetes providers; move it there.

KubernetesAccountResolver is really just a helper class for the
classes in .*kubernetes.v2.caching.view.provider; move it there
so it can be made package-private and only visible to those
classes.

GlobalKubernetesKindRegistry is closely associated with the
account credentials and should live in the same package (as
the account-specific kind registry does).
